refactor: refactor UserInterface and Validator
