+++ src/main/kotlin/string/calculator/CalculatorMachine.kt
@@ -3,11 +3,9 @@ package string.calculator
 import string.calculator.part.Calculator
 import string.calculator.part.Splitter
 import string.calculator.part.UserInterface
-import string.calculator.part.Validator
 
 class CalculatorMachine(
     private val calculator: Calculator,
-    private val validator: Validator,
     private val splitter: Splitter,
     private val userInterface: UserInterface
 ) {
@@ -25,7 +23,6 @@ class CalculatorMachine(
         while (isRunning) {
             val input = userInterface.introduce()
             val splitInput = splitter.split(input)
-            validator.validateInput(splitInput.numbers, splitInput.operators)
             val result = calculator.calculate(splitInput.numbers, splitInput.operators)
             userInterface.showResult(result)
             isRunning = userInterface.isContinue()
@@ -34,6 +31,6 @@ class CalculatorMachine(
 }
 
 fun main() {
-    val calculatorMachine = CalculatorMachine(Calculator, Validator(), Splitter(), UserInterface(Validator()))
+    val calculatorMachine = CalculatorMachine(Calculator, Splitter(), UserInterface)
     calculatorMachine.start()
 }
